requirements.txt
PyQt6
Pillow
bs4
ebooklib
recipe-scrapers
